[rename]파일이름 index에서 직접만든 이름으로 변경
