+++ src/pages/login/LoginPage.jsx
@@ -13,6 +13,7 @@ import { FillButton } from "../../components/Button";
 import kakao from "../../assets/login/Kakao.png";
 import naver from "../../assets/login/Naver.png";
 import { Link } from "react-router-dom";
+
 const Login = () => {
   const [isToggle, setIsToggle] = useState(false);
